Python binding is implemented.
Included objects OccurrencesFile and FileParser in Python binding so one can
read occurrences files just like in om_console.

Python SWIG is just "borrowing" those two objects for the time being. Those
object are outside the core om library.
We might want to have some kind of internal file reading in the core library
so it can be used by clients in a coherent way in the future

Also changed the name of method Occurrences::print() to ::printOccurrences()
to avoid colision with reserved word "print" in SWIG.