Made changes to OM::setAlgorithm(). It now creates the algorithm (_alg) and sampler objects (_samp).
Since those objects require other ones to exist (_presence and _env),
there is an order to call those OM methods now. The order is:

1) any of the following methods in any order:
  ::setOccurrences();
  ::setEnvironment();
  ::setOutput();

2) ::setAlgorithm();

3) ::run();

Changed the call order in om_console.cpp and test.py to reflect the changes
described above.

There is a SegmentationFault when OM object is deallocated. Will debug that
in Windows and commit changes in separate commit.