Changes in om_console to accomodate the introduction of the RequestFile class that is intended to set an OpenModeller object based on a request file.
The method OpenModeller::createMap was splited in the methods
setOutputMap() and createMap. This was done to let the map
properties be setted in a point different of the createMap()
calling. The createMap() continue to accept the "output file
name" and "output mask" as optional parameters.

For homogeneity reasons the "Output" request file tag was
changed to "Output file" and a the new tag "Output mask" was
created. So we have two masks tag: "Mask" and "Output mask".