Finished implementation of GARP with best subsets (as a separate algorithm). It runs GARP a number of time and "averages" the results. It links statically to GARP so there are no issues with algorithm dll versioning.
Added 1 public and 2 private methods to class Sampler to support
splitting of train and test occurrences. This might be modified later
depending on how ModelManager class handles test and train data.

I found an issue with WKT coordinate system names on the
Occurrences class constructor. Sometimes the only difference between
them are spaces in the definition of the coord system.
When OM compares them, it finds them different, but they really
aren't. I've added some C functions that might be used
in the comparison of names, but didn't make the modification
just yet. Take a look and let me know what you think.