The following serialization files need to be installed in "include/openmodeller/" installation directory:
* om_serializable.hh
* om_serializer_factory.hh
* om_deserializer.hh
* om_serializer.hh

The export of these header must be done because oM clients and/or
oM algorithms will need to include them. Think about an oM devel
installation.

Also some change were done so only <om.hh> needs to be included in
oM algorithms (in header, because they have to be derived from
Algorithm) and in oM clients.