Fixed a couple more memory leaks in SampledData _dep and _indep members. There still is a couple more die hard memory leaks there.
Changed ownership of memory allocated for Occurrences objects from OM object
to original client that allocated the memory.
That is because it is impossible for non-C++ external clients to transfer
ownership of allocated objects to the OpenModeller object. That is the
case for object Occurrences.
Once Java or Python create the Occurrences object, that language calls
the destructor of that object by default when they go out of scope,
thus deallocating its memory.
So in this case, Occurrences objects were deallocated twice, once by Java or
Python when the object went out of scope and once again when OpenModeller
was being destroyed, thus causing memory problems.
Changed om_console application accordingly, so it deallocates the Occurrences
object that it allocates in RequestFile::setOccurrences(...) in its destructor.
OpenModeller wont deallocate Occurrences objects (_presence and _absence)
anymore.