* Made significant changes to GARP algorithm:
1) Gapsize (number of rules kept from fittest individuals between iterations)
is now set to 10% (instead of 90%) just like in the previous versions of
GARP;

2) Changed name of methods Garp::select() to Garp::keepFittest() and
Garp::reproduce() to Garp::select() to conform with previous versions;

3) Modified some implementations of Rule::similar() methods (concrete
classes) to clean the code a bit. Also, changed LogitRule::similar()
to conform to former version. Changed NegatedRangeRule::similar() as
well;

4) Changed GARP unit tests accordinly.


* Changes to GARP_BS:

1) Fixed problems with interpretation of GARP_BS parameters for hard
omission threshold and calculation of omission and best subset of models.

* Other changes:

1) Fixed AccTest.py (user test scripts) to conform to changes in python
API (counted char * arrays and AlgMetadata.getParamList() method);

2) Added output of cell statistics and confusion matrix to console app;