Implemented basic abstraction to allow input from env layers + occurrences as well as from a bi-dimensional matrix of environmental values + occurrences.
In this implementation, Sampler is now an abstract class that supports
generic sampling services. The old Sampler became EnvironmentSampler which
is a concrete implementation of the base class. The 2nd concrete class
implemented is MatrixSampler which samples data from the bi-dimensional
matrix.

OM is still backward compatible, since calls to OM::setOccurrences() +
setEnvironment() will result in the creation of EnvironmentSampler that
will be passed on to setSampler().

Method OM::createMap is broken if a MatrixSampler() is passed to
OM::setSampler() because this method does not set either of the
OM data members _env, _presence nor _absence.