First set of major changes to support alternative Configuration based serialization/deserialization mechanisms.  At this point, one can extract the Configuration of an OpenModeller and write it in xml to disk.
-  All the algorithms which previously supported serialization have been
changed to support the new mechanism.  This list is:  garp, bioclim,
bioclim_distance, and distance_to_average.

-  Garp with best subsets had an empty serialization routine.  This has
been removed and will and will need to be written.

-  The scew library source in src/lib/serialization/scew has been
removed because the Configuration object is now our internal representation
of the configuration.  A custom event handler for expat was written
to deserialize the xml directly into Configuration.

-  The serialization code which was in OpenModeller has been distributed
to the appropriate object.  Before OpenModeller was responsible for
serializing the Environment, Occurrences, Algorithm Metadata and paramters.
Now that responsiblity is in the appropriate object.

-  Some deserialization stubs have been added.  In particular,
AlgorithmFactory, and Sampler (factory) have stubs added.