Added Model.hh/cpp abstract class for "Models".  A Model is the result of the execution of an Algorithm.  It provides the "getValue( Scalar * )" method which evaluates the species viability at a given environment vector.
Model implements a reference counted pointer to an implementation class.

Three concrete implementations of Model are given.  AlgoAdapterModel is
a short term adapter class which wraps an Algorithm in a Model interface.
ScaledModel is intended to be a base class for other models which do
environment scaling.  AverageModel takes the average viability from
a set of models (ie, Garp w/ best-subsets).

Algorithm was changed to implement Model.

Projector was changed to work only with Models and no longer with Algorithms.