First step in making Algorithm a reference counted pointer class which has reference semantics.  The Algorithm has been split into two classes Algorithm and AlgorithmImpl.
Each of the algorithm/* classes have been changed to derive from AlgorithmImpl.

Some effort has been made to clean up AlgorithmFactory.  In particular, a
"dll exposed" method has been added to each algorithm/* which returns a
pointer to the Algorithm's Metadata structure.  This means that
AlgorithmFactory does not need its own pointer to an algorithm.

There is one major problem outstanding.  In order to prevent
~AlgorithmFactory() from closing all the algorithm dll's prior to
OpenModeller destroying it's algorithm, OpenModeller::_algo is still
a pointer.  This will be changed shortly.