Add some const-ness to existing get routines.
Added ability to set the min/max values externally from the file i/o mechanism.
This was added so the min/max can be fixed by the configuration file so the
normalization does not need to be dependant on the masked region.

Change the implementation of calcMinMax to remove the return-value arguments.