The changes contained in this commit affect Windows build only.
Added macro "dllexp" to Occurrence, Header, Raster and ignorecase_traits classes so that they are accessible by clients across the DLL borders.

Changed VC7 build files so that they can find the source files on the new locations (and using the new names).
VC6 and VC8 builds are obsolete at this point.