Quite a few changes here: - created new method   gsl_matrix * CsmBS::randomiseColumns(gsl_matrix * original_matrix)   to separate out randomisation process. Randomiser now follows same technique as Neils   original matlab implementation whereby a random matrix is computed, permutations   are derived based on the sorted random cols and applied to the original input   matrix. - int CsmBS::discardComponents() now calls above method instead of doing it inline - cleaned up code formatting and indenting. Probably I should have done that as a separate   commit to make diffing easier :-(
The changes introduced mean that randomiser repeats are redundant and only the Randomisations
variable is used. I will remove this from parameters list in a future commit. Also stddev value of
2 should now be used - will set this as default in a future commit.

CSM is still a bit wobbly so not ready to be used properly yet until I resolve a last few
(hopefully small) issues.